+++ packages/lib/test/model/valueType.test.ts
@@ -1,3 +1,4 @@
+import { set } from "mobx"
 import { Model, model, modelAction, prop } from "../../src"
 
 test("value type", () => {
@@ -41,7 +42,8 @@ test("value type", () => {
 
     @modelAction
     objSet(p: P | undefined) {
-      this.p_obj.p = p
+      // this.p_obj.p = p // not supported by mobx4
+      set(this.p_obj, "p", p)
       return this.p_obj.p
     }
   }