+++ packages/lib/test/snapshot/modelProcessor.test.ts
@@ -1,3 +1,4 @@
+import { toJS } from "mobx"
 import { assert, _ } from "spec.ts"
 import {
   applyPatches,
@@ -55,7 +56,7 @@ test("input snapshot processor", () => {
     })
   )
 
-  expect(p.arr).toEqual([30, 40, 50])
+  expect(toJS(p.arr)).toEqual([30, 40, 50])
 
   applyPatches(p, [
     {
@@ -65,7 +66,7 @@ test("input snapshot processor", () => {
     },
   ])
 
-  expect(p.arr).toEqual([10, 20])
+  expect(toJS(p.arr)).toEqual([10, 20])
 
   applySnapshot(
     p,
@@ -74,11 +75,11 @@ test("input snapshot processor", () => {
     })
   )
 
-  expect(p.arr).toEqual([100, 200])
+  expect(toJS(p.arr)).toEqual([100, 200])
 
   applySnapshot(p.arr, [300, 400])
 
-  expect(p.arr).toEqual([300, 400])
+  expect(toJS(p.arr)).toEqual([300, 400])
 })
 
 test("input snapshot processor with original type", () => {
@@ -124,7 +125,7 @@ test("input snapshot processor with original type", () => {
     })
   )
 
-  expect(p.arr).toEqual([30, 40, 50])
+  expect(toJS(p.arr)).toEqual([30, 40, 50])
 
   applyPatches(p, [
     {
@@ -134,7 +135,7 @@ test("input snapshot processor with original type", () => {
     },
   ])
 
-  expect(p.arr).toEqual([10, 20])
+  expect(toJS(p.arr)).toEqual([10, 20])
 
   applySnapshot(
     p,
@@ -143,11 +144,11 @@ test("input snapshot processor with original type", () => {
     })
   )
 
-  expect(p.arr).toEqual([100, 200])
+  expect(toJS(p.arr)).toEqual([100, 200])
 
   applySnapshot(p.arr, [300, 400])
 
-  expect(p.arr).toEqual([300, 400])
+  expect(toJS(p.arr)).toEqual([300, 400])
 })
 
 test("output snapshot processor", () => {