+++ packages/lib/test/snapshot/propProcessor.test.ts
@@ -1,3 +1,4 @@
+import { toJS } from "mobx"
 import { assert, _ } from "spec.ts"
 import {
   applyPatches,
@@ -45,7 +46,7 @@ test("input snapshot processor", () => {
     arr: [30, 40, 50],
   })
 
-  expect(p2.arr).toEqual([30, 40, 50])
+  expect(toJS(p2.arr)).toEqual([30, 40, 50])
 
   const p = fromSnapshot<P3>(
     modelSnapshotInWithMetadata(P3, {
@@ -53,7 +54,7 @@ test("input snapshot processor", () => {
     })
   )
 
-  expect(p.arr).toEqual([30, 40, 50])
+  expect(toJS(p.arr)).toEqual([30, 40, 50])
 
   applyPatches(p, [
     {
@@ -63,7 +64,7 @@ test("input snapshot processor", () => {
     },
   ])
 
-  expect(p.arr).toEqual([10, 20])
+  expect(toJS(p.arr)).toEqual([10, 20])
 
   applySnapshot(
     p,
@@ -72,11 +73,11 @@ test("input snapshot processor", () => {
     })
   )
 
-  expect(p.arr).toEqual([100, 200])
+  expect(toJS(p.arr)).toEqual([100, 200])
 
   applySnapshot(p.arr, [300, 400])
 
-  expect(p.arr).toEqual([300, 400])
+  expect(toJS(p.arr)).toEqual([300, 400])
 })
 
 test("output snapshot processor", () => {